修改

+++ exercises/enums/enums2.rs
@@ -5,10 +5,9 @@
 
 #[derive(Debug)]
 enum Message {
-    // TODO: define the different variants used below
     Move { x: i32, y: i32 },
     Echo(String),
-    ChangeColor(i32, i32, i32),
+    ChangeColor(u8, u8, u8),
     Quit,
 }
